+++ src/main/java/com/moa/moa_server/config/jpa/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.moa.moa_server.config;
+package com.moa.moa_server.config.jpa;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
